Started Input refactor * Renamed input_base module to input * Added ABC for inputs * Check ABC rather than existence of "set_input"
