+++ ml_genn/ml_genn/neurons/__init__.py
@@ -2,6 +2,7 @@ from .adaptive_leaky_integrate_fire import AdaptiveLeakyIntegrateFire
 from .binary_spike_input import BinarySpikeInput
 from .few_spike_relu import FewSpikeRelu
 from .few_spike_relu_input import FewSpikeReluInput
+from .input import Input
 from .integrate_fire import IntegrateFire
 from .integrate_fire_input import IntegrateFireInput
 from .leaky_integrate import LeakyIntegrate
@@ -15,6 +16,6 @@ from ..utils.module import get_module_classes
 default_neurons = get_module_classes(globals(), Neuron)
 
 __all = ["AdaptiveLeakyIntegrateFire", "BinarySpikeInput", "FewSpikeRelu",
-         "FewSpikeReluInput", "IntegrateFire", "IntegrateFireInput",
+         "FewSpikeReluInput", "Input", "IntegrateFire", "IntegrateFireInput",
          "LeakyIntegrate", "LeakyIntegrateFire", "Neuron",
          "PoissonInput", "SpikeInput", "default_neurons"]
