+++ ml_genn/ml_genn/neurons/integrate_fire_input.py
@@ -1,5 +1,5 @@
 from pygenn.genn_wrapper.Models import VarAccess_READ_ONLY_DUPLICATE
-from .input_base import InputBase
+from .input import InputBase
 from .neuron import Neuron
 from ..utils.model import NeuronModel
 from ..utils.value import InitValue, ValueDescriptor
